style : rename utilitites -> utilities
